contract interaction fixing
